没有生产权限的账户,隐藏构建列表中的记录

+++ application/clusterApp/app.go
@@ -4,6 +4,7 @@ package clusterApp
 import (
 	"fops/application/clusterApp/request"
 	"fops/domain/cluster"
+
 	"github.com/farseer-go/collections"
 	"github.com/farseer-go/fs/exception"
 	"github.com/farseer-go/mapper"
@@ -49,7 +50,7 @@ func List(clusterRepository cluster.Repository) collections.List[cluster.DomainO
 // Delete 删除集群
 // @post delete
 // @filter application.Jwt
-func Delete(clusterId int64, clusterRepository cluster.Repository) {
+func Delete(clusterId int, clusterRepository cluster.Repository) {
 	exception.ThrowWebExceptionBool(clusterId < 1, 403, "集群Id没有填")
 	_, err := clusterRepository.Delete(clusterId)
 	exception.ThrowWebExceptionError(403, err)
