+++ domain/apps/repository.go
@@ -12,8 +12,8 @@ import (
 type Repository interface {
 	data.IRepository[DomainObject] // IRepository 通用的仓储接口
 	UpdateApp(do DomainObject) error
-	UpdateClusterVer(appName string, dicClusterVer collections.Dictionary[int64, ClusterVerVO]) (int64, error) // UpdateClusterVer 修改集群的镜像版本
-	UpdateInspect(lst collections.List[DomainObject]) (int64, error)                                           // UpdateInspect 更新从集群中获取到的实例、副本数量
+	UpdateClusterVer(appName string, dicClusterVer collections.Dictionary[int, ClusterVerVO]) (int64, error) // UpdateClusterVer 修改集群的镜像版本
+	UpdateInspect(lst collections.List[DomainObject]) (int64, error)                                         // UpdateInspect 更新从集群中获取到的实例、副本数量
 	ToListBySys(isSys bool) collections.List[DomainObject]
 	ToShortList(isAll bool) collections.List[ShortEO]
 	ToUTList() collections.List[DomainObject] // 获取需要UT的应用
